Refactor status reporter
